+++ ImageUtils.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+exec java -cp nikiroo-utils.jar be.nikiroo.utils.main.StartImageUtils "$@"
+