+++ src/be/nikiroo/utils/main/StartStringUtils.java
@@ -1,0 +1,5 @@
+package be.nikiroo.utils.main;
+
+public class StartStringUtils {
+
+}